Revert "Fix signed/unsigned comparison"
Unfortunately using unsigned ints isn't a solution here, because we are
creating negative numbers in this code which is UB then.

This reverts commit 92b3ec5fb6d2781c83edc488a4b72fab918b707d.

+++ tests/test-output-flex-way-del.cpp
@@ -31,8 +31,8 @@ TEST_CASE("delete way: not a member")
 
     testing::db::data_t data{tdata};
 
-    unsigned long num_t1 = 0;
-    unsigned long num_tboth = 0;
+    int num_t1 = 0;
+    int num_tboth = 0;
     SECTION("in none") { data.add("w10 v1 dV Tt=ag Nn10,n11"); }
     SECTION("in t1")
     {
@@ -69,9 +69,9 @@ TEST_CASE("delete way: relation member")
 
     testing::db::data_t data{tdata};
 
-    unsigned long num_t1 = 0;
-    unsigned long num_t2 = 0;
-    unsigned long num_tboth = 0;
+    int num_t1 = 0;
+    int num_t2 = 0;
+    int num_tboth = 0;
 
     SECTION("in none")
     {
